부족한 금액 계산

+++ src/cal/CalSolution.java
@@ -7,7 +7,7 @@ public class CalSolution {
 
 	public long solution(int price, int money, int count) {
 		long answer = -1;
-		answer = (long) price * count * (count + 1) / 2 - money;
+		answer = price * count * (count + 1) / 2 - money;
 		return answer <= 0 ? 0 : answer;
 	}
 }
